requirements.txt
numpy
matplotlib
pillow
warnings
